fix(events): use int for reviewID in StripRetiredDupWarnings interface
- Change interface signature from int32 to int (matching all sibling methods)
- Move int32 cast into EventRepository implementation
- Add stripRetiredDupWarningsFunc mocks to companion replacement tests
- Update all 4 mock files to use int

Generated-by: deepseek-v4-pro

+++ internal/domain/events/admin_service.go
@@ -1892,7 +1892,7 @@ func (s *AdminService) consolidateStripRetiredDupWarnings(
 		return dismissedIDs, nil
 	}
 
-	warningsEmpty, err := txRepo.StripRetiredDupWarnings(ctx, int32(entry.ID), retireULIDs)
+	warningsEmpty, err := txRepo.StripRetiredDupWarnings(ctx, entry.ID, retireULIDs)
 	if err != nil {
 		return dismissedIDs, fmt.Errorf("strip retired dup warnings for review %d: %w", entry.ID, err)
 	}
